refactor(rac): drop compatibility shims; import layers directly [roadmap:v0.7.4]
The v0.7.4 layering shipped thin re-export shims at the old top-level module
paths (rac/parser.py -> rac.core.markdown, rac/outputs.py -> rac.output, etc.)
to keep tests untouched. They read as duplicate modules in the source tree and
add no value now that the layer packages are the canonical API.

- Repoint test imports to rac.core.* / rac.services.* / rac.output.
- Remove the 15 src/rac/*.py compatibility shims.

No behavior change; CLI and JSON contracts unchanged.
