--- src/rac/artifacts.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.artifacts` (v0.7.4)."""
-
-from rac.core.artifacts import *  # noqa: F401,F403
