--- src/rac/classification.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.classification` (v0.7.4)."""
-
-from rac.core.classification import *  # noqa: F401,F403
