--- src/rac/diff.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.diff` (v0.7.4)."""
-
-from rac.services.diff import *  # noqa: F401,F403
