--- src/rac/fs.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.fs` (v0.7.4)."""
-
-from rac.core.fs import *  # noqa: F401,F403
