--- src/rac/improve.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.improve` (v0.7.4)."""
-
-from rac.services.improve import *  # noqa: F401,F403
