--- src/rac/inspect.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.inspect` (v0.7.4)."""
-
-from rac.services.inspect import *  # noqa: F401,F403
