--- src/rac/models.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.models` (v0.7.4)."""
-
-from rac.core.models import *  # noqa: F401,F403
