--- src/rac/parser.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.markdown` (v0.7.4)."""
-
-from rac.core.markdown import *  # noqa: F401,F403
