--- src/rac/portfolio.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.portfolio` (v0.7.4)."""
-
-from rac.services.portfolio import *  # noqa: F401,F403
