--- src/rac/relationships.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.relationships` (v0.7.4)."""
-
-from rac.services.relationships import *  # noqa: F401,F403
