--- src/rac/schema.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.schema` (v0.7.4)."""
-
-from rac.core.schema import *  # noqa: F401,F403
