--- src/rac/stats.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.services.stats` (v0.7.4)."""
-
-from rac.services.stats import *  # noqa: F401,F403
