--- src/rac/validate.py
@@ -1,3 +1,0 @@
-"""Compatibility shim — implementation moved to :mod:`rac.core.validation` (v0.7.4)."""
-
-from rac.core.validation import *  # noqa: F401,F403
